[#1733] Adds a new table `form_instance_download`
A less intrusive change would be to add a `datapoint_id` to the
existing `data_point_download` table, but we need to remove the `NOT
NULL` property of the existing `survey_id` column. It seems that is
**not** possible to do it in sqlite. You need to copy the data over,
recreate the table and copy back. Too much risk.

We opted for creating a new table instead: `form_instance_download`

Co-authored-by: Daniel Sunnerek <daniel@akvo.org>